Refactoring, with fix for incorrect uri passed.
