+++ src/main/java/serverlets/OAuthLogin.java
@@ -1,5 +1,7 @@
-package beans;
+package serverlets;
 
+import beans.EnvironmentVariableClass;
+import beans.OAuthBean;
 import com.github.scribejava.apis.fitbit.FitBitOAuth2AccessToken;
 import com.github.scribejava.core.model.OAuth2AccessToken;
 import persistence.TokenMap;
